+++ fluid/PaddleCV/image_classification/models/shufflenet_v2.py
@@ -67,7 +67,6 @@ class ShuffleNetV2():
 
         output = fluid.layers.fc(input=pool_last,
                                  size=class_dim,
-                                 act='softmax',
                                  param_attr=ParamAttr(initializer=MSRA(),name='fc6_weights'),
                                  bias_attr=ParamAttr(name='fc6_offset'))
         return output
