+++ fluid/PaddleCV/image_classification/reader.py
@@ -4,8 +4,7 @@ import random
 import functools
 import numpy as np
 import paddle
-import cv2
-import io
+from PIL import Image, ImageEnhance
 
 random.seed(0)
 np.random.seed(0)
@@ -16,64 +15,22 @@ THREAD = 8
 BUF_SIZE = 102400
 
 DATA_DIR = 'data/ILSVRC2012'
+DATA_DIR = '/work/data/cv'
+
 img_mean = np.array([0.485, 0.456, 0.406]).reshape((3, 1, 1))
 img_std = np.array([0.229, 0.224, 0.225]).reshape((3, 1, 1))
 
 
-def rotate_image(img):
-    """ rotate_image """
-    (h, w) = img.shape[:2]
-    center = (w / 2, h / 2)
-    angle = np.random.randint(-10, 11)
-    M = cv2.getRotationMatrix2D(center, angle, 1.0)
-    rotated = cv2.warpAffine(img, M, (w, h))
-    return rotated
-
-
-def random_crop(img, size, scale=None, ratio=None):
-    """ random_crop """
-    scale = [0.08, 1.0] if scale is None else scale
-    ratio = [3. / 4., 4. / 3.] if ratio is None else ratio
-
-    aspect_ratio = math.sqrt(np.random.uniform(*ratio))
-    w = 1. * aspect_ratio
-    h = 1. / aspect_ratio
-
-
-    bound = min((float(img.shape[0]) / img.shape[1]) / (w**2),
-                (float(img.shape[1]) / img.shape[0]) / (h**2))
-    scale_max = min(scale[1], bound)
-    scale_min = min(scale[0], bound)
-
-    target_area = img.shape[0] * img.shape[1] * np.random.uniform(scale_min,
-                                                                  scale_max)
-    target_size = math.sqrt(target_area)
-    w = int(target_size * w)
-    h = int(target_size * h)
-    i = np.random.randint(0, img.shape[0] - w + 1)
-    j = np.random.randint(0, img.shape[1] - h + 1)
-
-    img = img[i:i + w, j:j + h, :]
-
-    resized = cv2.resize(img, (size, size), interpolation=cv2.INTER_LANCZOS4)
-    return resized
-
-def distort_color(img):
+def resize_short(img, target_size):
+    percent = float(target_size) / min(img.size[0], img.size[1])
+    resized_width = int(round(img.size[0] * percent))
+    resized_height = int(round(img.size[1] * percent))
+    img = img.resize((resized_width, resized_height), Image.LANCZOS)
     return img
 
 
-def resize_short(img, target_size):
-    """ resize_short """
-    percent = float(target_size) / min(img.shape[0], img.shape[1])
-    resized_width = int(round(img.shape[1] * percent))
-    resized_height = int(round(img.shape[0] * percent))
-    resized = cv2.resize(img, (resized_width, resized_height), interpolation=cv2.INTER_LANCZOS4)
-    return resized
-
-
 def crop_image(img, target_size, center):
-    """ crop_image """
-    height, width = img.shape[:2]
+    width, height = img.size
     size = target_size
     if center == True:
         w_start = (width - size) / 2
@@ -83,55 +40,90 @@ def crop_image(img, target_size, center):
         h_start = np.random.randint(0, height - size + 1)
     w_end = w_start + size
     h_end = h_start + size
-    img = img[h_start:h_end, w_start:w_end, :]
+    img = img.crop((w_start, h_start, w_end, h_end))
     return img
 
 
-def process_image(sample,
-                  mode,
-                  color_jitter,
-                  rotate,
-                  crop_size=224,
-                  mean=None,
-                  std=None):
-    """ process_image """
+def random_crop(img, size, scale=[0.08, 1.0], ratio=[3. / 4., 4. / 3.]):
+    aspect_ratio = math.sqrt(np.random.uniform(*ratio))
+    w = 1. * aspect_ratio
+    h = 1. / aspect_ratio
 
-    mean = [0.485, 0.456, 0.406] if mean is None else mean
-    std = [0.229, 0.224, 0.225] if std is None else std
+    bound = min((float(img.size[0]) / img.size[1]) / (w**2),
+                (float(img.size[1]) / img.size[0]) / (h**2))
+    scale_max = min(scale[1], bound)
+    scale_min = min(scale[0], bound)
 
+    target_area = img.size[0] * img.size[1] * np.random.uniform(scale_min,
+                                                                scale_max)
+    target_size = math.sqrt(target_area)
+    w = int(target_size * w)
+    h = int(target_size * h)
+
+    i = np.random.randint(0, img.size[0] - w + 1)
+    j = np.random.randint(0, img.size[1] - h + 1)
+
+    img = img.crop((i, j, i + w, j + h))
+    img = img.resize((size, size), Image.LANCZOS)
+    return img
+
+
+def rotate_image(img):
+    angle = np.random.randint(-10, 11)
+    img = img.rotate(angle)
+    return img
+
+
+def distort_color(img):
+    def random_brightness(img, lower=0.5, upper=1.5):
+        e = np.random.uniform(lower, upper)
+        return ImageEnhance.Brightness(img).enhance(e)
+
+    def random_contrast(img, lower=0.5, upper=1.5):
+        e = np.random.uniform(lower, upper)
+        return ImageEnhance.Contrast(img).enhance(e)
+
+    def random_color(img, lower=0.5, upper=1.5):
+        e = np.random.uniform(lower, upper)
+        return ImageEnhance.Color(img).enhance(e)
+
+    ops = [random_brightness, random_contrast, random_color]
+    np.random.shuffle(ops)
+
+    img = ops[0](img)
+    img = ops[1](img)
+    img = ops[2](img)
+
+    return img
+
+
+def process_image(sample, mode, color_jitter, rotate):
     img_path = sample[0]
-    img = cv2.imread(img_path)
 
+    img = Image.open(img_path)
     if mode == 'train':
-        if rotate:
-            img = rotate_image(img)
-        if crop_size > 0:
-            img = random_crop(img, crop_size)
+        if rotate: img = rotate_image(img)
+        img = random_crop(img, DATA_DIM)
+    else:
+        img = resize_short(img, target_size=256)
+        img = crop_image(img, target_size=DATA_DIM, center=True)
+    if mode == 'train':
         if color_jitter:
             img = distort_color(img)
         if np.random.randint(0, 2) == 1:
-            img = img[:, ::-1, :]
-    else:
-        if crop_size > 0:
-            img = resize_short(img, crop_size)
+            img = img.transpose(Image.FLIP_LEFT_RIGHT)
 
-            img = crop_image(img, target_size=crop_size, center=True)
+    if img.mode != 'RGB':
+        img = img.convert('RGB')
 
-    img = img[:, :, ::-1].astype('float32').transpose((2, 0, 1)) / 255
-    img_mean = np.array(mean).reshape((3, 1, 1))
-    img_std = np.array(std).reshape((3, 1, 1))
+    img = np.array(img).astype('float32').transpose((2, 0, 1)) / 255
     img -= img_mean
     img /= img_std
 
     if mode == 'train' or mode == 'val':
-        return (img, sample[1])
+        return img, sample[1]
     elif mode == 'test':
-        return (img, )
-
-
-def image_mapper(**kwargs):
-    """ image_mapper """
-    return functools.partial(process_image, **kwargs)
+        return [img]
 
 
 def _reader_creator(file_list,
@@ -165,47 +157,44 @@ def _reader_creator(file_list,
             for line in lines:
                 if mode == 'train' or mode == 'val':
                     img_path, label = line.split()
-                    img_path = img_path.replace("JPEG", "jpeg")
+                    #img_path = img_path.replace("JPEG", "jpeg")
                     img_path = os.path.join(data_dir, img_path)
                     yield img_path, int(label)
                 elif mode == 'test':
                     img_path, label = line.split()
-                    img_path = img_path.replace("JPEG", "jpeg")
+                    #img_path = img_path.replace("JPEG", "jpeg")
                     img_path = os.path.join(data_dir, img_path)
- 
+
                     yield [img_path]
 
-    image_mapper = functools.partial(
-        process_image,
-        mode=mode,
-        color_jitter=color_jitter,
-        rotate=rotate,
-        crop_size=224)
-    reader = paddle.reader.xmap_readers(
-        image_mapper, reader, THREAD, BUF_SIZE, order=False)
-    return reader
+    mapper = functools.partial(
+        process_image, mode=mode, color_jitter=color_jitter, rotate=rotate)
+
+    return paddle.reader.xmap_readers(mapper, reader, THREAD, BUF_SIZE)
 
 
 def train(data_dir=DATA_DIR, pass_id_as_seed=0):
-
-    file_list = os.path.join(data_dir, 'dataset_100/train_list.txt')
+    file_list = os.path.join(data_dir, 'train_list.txt')
     return _reader_creator(
         file_list,
         'train',
         shuffle=True,
         color_jitter=False,
         rotate=False,
-        data_dir=data_dir,
+        #data_dir=data_dir,
+        data_dir=os.path.join(data_dir,'train'),
         pass_id_as_seed=pass_id_as_seed)
 
 
-def val(data_dir=DATA_DIR): 
+def val(data_dir=DATA_DIR):
     file_list = os.path.join(data_dir, 'val_list.txt')
     return _reader_creator(file_list, 'val', shuffle=False, 
-            data_dir=data_dir)
+            data_dir=os.path.join(data_dir,'val'))
+         #   data_dir=data_dir)
 
 
 def test(data_dir=DATA_DIR):
     file_list = os.path.join(data_dir, 'val_list.txt')
-    return _reader_creator(file_list, 'test', shuffle=False,
-            data_dir=data_dir)
+    return _reader_creator(file_list, 'test', shuffle=False, 
+            data_dir=os.path.join(data_dir,'val'))
+      #      data_dir=data_dir)
